add official repository functionality and validation
