+++ app/templates/repositories/repository_form.html
@@ -20,7 +20,7 @@
 
                         <div class="mb-4">
                             <label for="{{ form.name.id_for_label }}" class="form-label fw-semibold">
-                                Repository Name <span class="text-danger">*</span>
+                                Name <span class="text-danger">*</span>
                             </label>
                             <div class="input-group">
                                 <span class="input-group-text">
@@ -71,6 +71,15 @@
                                 <strong>Private:</strong> Only you can see this repository
                             </div>
                         </div>
+
+                        {% if user.is_staff %}
+                        <div class="mb-2">
+                            <label for="{{ form.is_official.id_for_label }}" class="form-label fw-semibold">
+                                Official <span class="text-danger">*</span>
+                            </label>
+                            <input type="checkbox" class="form-check-input" id="{{ form.is_official.id_for_label }}" name="{{ form.is_official.html_name }}" {% if form.is_official.value %}checked{% endif %}/>
+                        </div>
+                        {% endif %}
 
                         <div class="d-flex gap-2 pt-3 border-top">
                             <button type="submit" class="btn btn-success">
